add: 数据库实体重新生成&用户登录&权限体系

+++ webserver/src/main/java/cn/qkl/webserver/vo/user/UserInfoVO.java
@@ -19,7 +19,7 @@ public class UserInfoVO {
     private Long userId;
 
     @ApiModelProperty("角色")
-    private Byte roleType;
+    private Integer roleType;
 
     @ApiModelProperty("昵称")
     private String nickName;
